Move `EqOp` into `Operators` lint pass
