+++ tests/ui/zero_div_zero.stderr
@@ -1,11 +1,3 @@
-error: equal expressions as operands to `/`
-  --> $DIR/zero_div_zero.rs:4:15
-   |
-LL |     let nan = 0.0 / 0.0;
-   |               ^^^^^^^^^
-   |
-   = note: `#[deny(clippy::eq_op)]` on by default
-
 error: constant division of `0.0` with `0.0` will always result in NaN
   --> $DIR/zero_div_zero.rs:4:15
    |
@@ -15,12 +7,6 @@ LL |     let nan = 0.0 / 0.0;
    = note: `-D clippy::zero-divided-by-zero` implied by `-D warnings`
    = help: consider using `f64::NAN` if you would like a constant representing NaN
 
-error: equal expressions as operands to `/`
-  --> $DIR/zero_div_zero.rs:5:19
-   |
-LL |     let f64_nan = 0.0 / 0.0f64;
-   |                   ^^^^^^^^^^^^
-
 error: constant division of `0.0` with `0.0` will always result in NaN
   --> $DIR/zero_div_zero.rs:5:19
    |
@@ -28,12 +14,6 @@ LL |     let f64_nan = 0.0 / 0.0f64;
    |                   ^^^^^^^^^^^^
    |
    = help: consider using `f64::NAN` if you would like a constant representing NaN
-
-error: equal expressions as operands to `/`
-  --> $DIR/zero_div_zero.rs:6:25
-   |
-LL |     let other_f64_nan = 0.0f64 / 0.0;
-   |                         ^^^^^^^^^^^^
 
 error: constant division of `0.0` with `0.0` will always result in NaN
   --> $DIR/zero_div_zero.rs:6:25
@@ -43,12 +23,6 @@ LL |     let other_f64_nan = 0.0f64 / 0.0;
    |
    = help: consider using `f64::NAN` if you would like a constant representing NaN
 
-error: equal expressions as operands to `/`
-  --> $DIR/zero_div_zero.rs:7:28
-   |
-LL |     let one_more_f64_nan = 0.0f64 / 0.0f64;
-   |                            ^^^^^^^^^^^^^^^
-
 error: constant division of `0.0` with `0.0` will always result in NaN
   --> $DIR/zero_div_zero.rs:7:28
    |
@@ -57,5 +31,5 @@ LL |     let one_more_f64_nan = 0.0f64 / 0.0f64;
    |
    = help: consider using `f64::NAN` if you would like a constant representing NaN
 
-error: aborting due to 8 previous errors
+error: aborting due to 4 previous errors
